+++ packages/database/src/api/onDisconnect.ts
@@ -17,9 +17,9 @@
 
 import { validateArgCount, validateCallback, Compat } from '@firebase/util';
 
+import { OnDisconnect as ExpOnDisconnect } from '../../exp/index';
 import { Indexable } from '../core/util/misc';
 import { warn } from '../core/util/util';
-import { OnDisconnect as ExpOnDisconnect } from '../exp/OnDisconnect';
 
 export class OnDisconnect implements Compat<ExpOnDisconnect> {
   constructor(readonly _delegate: ExpOnDisconnect) {}